requirements.txt
fastapi
uvicorn[standard]
websockets
sqlalchemy
psycopg[binary]
alembic
pydantic-settings
bcrypt
PyJWT
python-multipart
boto3